# Adds cudnn feature flag. Removes "test-cuda" feature flag. Using cuDNN for convolutions