--- a/src/tensor/cuda/device.rs
+++ b/src/tensor/cuda/device.rs
@@ -2,6 +2,8 @@ use crate::shapes::{Shape, Unit};
 use crate::tensor::cpu::{Cpu, CpuError, NdIndex};
 use crate::tensor::{DeviceStorage, HasErr, Tensor};
 
+use cudarc::cudnn::result::CudnnError;
+use cudarc::cudnn::Cudnn;
 use cudarc::{
     cublas::{result::CublasError, CudaBlas},
     driver::{CudaDevice, CudaSlice, CudaStream, DeviceSlice, DriverError},
@@ -20,6 +22,7 @@ pub struct Cuda {
     pub(crate) cpu: Cpu,
     pub(crate) dev: Arc<CudaDevice>,
     pub(crate) blas: Arc<CudaBlas>,
+    pub(crate) cudnn: Arc<Cudnn>,
     /// A second stream for kernels to optionally execute on.
     pub(crate) par_stream: Arc<CudaStream>,
     pub(crate) workspace: Arc<Mutex<CudaSlice<u8>>>,
@@ -28,6 +31,7 @@ pub struct Cuda {
 #[derive(Debug)]
 pub enum CudaError {
     Blas(CublasError),
+    Cudnn(CudnnError),
     Driver(DriverError),
     Cpu(CpuError),
 }
@@ -47,6 +51,12 @@ impl From<CublasError> for CudaError {
 impl From<DriverError> for CudaError {
     fn from(value: DriverError) -> Self {
         Self::Driver(value)
+    }
+}
+
+impl From<CudnnError> for CudaError {
+    fn from(value: CudnnError) -> Self {
+        Self::Cudnn(value)
     }
 }
 
@@ -72,12 +82,14 @@ impl Cuda {
         let cpu = Cpu::seed_from_u64(seed);
         let dev = CudaDevice::new(ordinal)?;
         let blas = Arc::new(CudaBlas::new(dev.clone())?);
+        let cudnn = Arc::new(Cudnn::new(dev.clone())?);
         let par_stream = Arc::new(dev.fork_default_stream()?);
         let workspace = Arc::new(Mutex::new(dev.alloc_zeros::<u8>(0)?));
         Ok(Self {
             cpu,
             dev,
             blas,
+            cudnn,
             par_stream,
             workspace,
         })