Resolves #638
Resolves #527 

TODOS
- Should we optionally enable cudnn?

Timings of `cargo +nightly bench --bench conv2d -F cuda` on an A10

| branch | fwd | bwd |
| --- | --- | --- |
| main | 4.73ms | 20ms |
| this | 2.6ms | 8ms |