@opfromthestart 2023-04-01
How does this compare with the old implementation on the conv2d benchmark?