@coreylowman 2023-04-02
@opfromthestart updated PR description with timings, its much faster 😀